'linear' burn function
