+++ packages/vm/gas/table.go
@@ -1,5 +1,7 @@
 package gas
 
+import "golang.org/x/xerrors"
+
 const (
 	BurnStorage1P = BurnCode(iota)
 	BurnCallTargetNotFound
@@ -30,8 +32,8 @@ const (
 	BurnWasm1P
 )
 
+// burnTable contains all possible burn codes with their burn value computing functions
 var burnTable = BurnTable{
-	// constant values
 	BurnCallTargetNotFound:         {"target n/f", constValue(10)},
 	BurnGetContext:                 {"context", constValue(10)},
 	BurnGetCallerData:              {"caller", constValue(10)},
@@ -43,8 +45,8 @@ var burnTable = BurnTable{
 	BurnTransferAllowance:          {"transfer", constValue(10)},
 	BurnSendL1Request:              {"send", constValue(10)},
 	BurnDeployContract:             {"deploy", constValue(10)},
-	BurnStorage1P:                  {"storage", proportion()},
-	BurnWasm1P:                     {"storage", proportion()},
+	BurnStorage1P:                  {"storage", linear()},
+	BurnWasm1P:                     {"wasm", linear()},
 	BurnUtilsHashingBlake2b:        {"blake2b", constValue(50)},
 	BurnUtilsHashingSha3:           {"sha3", constValue(80)},
 	BurnUtilsHashingHname:          {"hname", constValue(50)},
@@ -53,29 +55,67 @@ var burnTable = BurnTable{
 	BurnUtilsED25519ValidSig:       {"ed25517 valid", constValue(200)},
 	BurnUtilsED25519AddrFromPubKey: {"ed25517 addr", constValue(50)},
 	BurnUtilsBLSValidSignature:     {"bls valid", constValue(2000)},
-	BurnUtilsBLSAddrFromPubKey:     {"ed25517 addr", constValue(50)},
-	BurnUtilsBLSAggregateBLS1P:     {"bls aggregate", proportion(400)},
+	BurnUtilsBLSAddrFromPubKey:     {"bls addr", constValue(50)},
+	BurnUtilsBLSAggregateBLS1P:     {"bls aggregate", linear(400)},
 }
 
-// proportion construct a closure which returns gas proportional to the first parameter
-// if proportion() without parameters return close which return gas equal to the first parameter
-func proportion(p ...int) BurnFunction {
-	if len(p) == 0 {
-		return func(_ BurnCode, params []int) uint64 {
-			if len(params) != 1 {
-				panic("'proportion' burn function requires exactly 1 parameter")
+func constValue(constGas uint64) BurnFunction {
+	g := constGas
+	return func(_ BurnCode, _ []int) uint64 {
+		return g
+	}
+}
+
+func notImplemented() BurnFunction {
+	return func(code BurnCode, _ []int) uint64 {
+		panic(xerrors.Errorf("burn code %d not implemented", code))
+	}
+}
+
+func (c BurnCode) Value(p ...int) uint64 {
+	return Value(c, p...)
+}
+
+// linear takes A and B as parameters ans construct a closure which returns gas as a linear function A*x+B of the parameter x.
+// linear() without parameters returns closure corresponding to identity function 1*x+0
+func linear(p ...int) BurnFunction {
+	switch len(p) {
+	case 0:
+		// burn value == X
+		return func(_ BurnCode, x []int) uint64 {
+			if len(x) != 1 {
+				panic(ErrInLinear1ParameterExpected)
 			}
-			return uint64(params[0])
+			return uint64(x[0])
 		}
-	}
-	if p[0] <= 0 {
-		panic("proportion: wrong parameter")
-	}
-	c := uint64(p[0])
-	return func(_ BurnCode, params []int) uint64 {
-		if len(params) != 1 {
-			panic("'proportion' burn function requires exactly 1 parameter")
+	case 1:
+		a := uint64(p[0])
+		// burn value == a*X
+		return func(_ BurnCode, x []int) uint64 {
+			if len(x) != 1 {
+				panic(ErrInLinear1ParameterExpected)
+			}
+			return a * uint64(x[0])
 		}
-		return c * uint64(params[0])
+	case 2:
+		a := uint64(p[0])
+		b := uint64(p[1])
+		if a == 0 {
+			// burn value = 0*X+b
+			return constValue(b)
+		}
+		if b == 0 {
+			// burn value = a*X
+			return linear(p[0])
+		}
+		// burn value = a*X+b
+		return func(_ BurnCode, x []int) uint64 {
+			if len(x) != 1 {
+				panic(ErrInLinear1ParameterExpected)
+			}
+			return a*uint64(x[0]) + b
+		}
+	default:
+		panic("function requires 0, 1 or 2 parameter")
 	}
 }
